AML: DVB: fix for SC2 DVB not receiving stream from demod
background:
Demux driver does not work on SC2 device that was originally
released with kernel v4.9 (mecool KT1)
The issue was introduced in kernel v5.4 with:

commit 15954fc01f8a97f897d6daf21699a11f466e8d1a
Author: chuangcheng peng <chuangcheng.peng@amlogic.com>
Date:   Thu Mar 3 19:16:01 2022 +0800

    demux: use new tee interface to set [1/1]

    PD#SWPL-74508

    Problem:
    tee flag conflict

    Solution:
    call new tee interface

    Verify:
    not verified

    Change-Id: Iccd8988af3a2ec612d858951dfc871d97bb4f1ac
    Signed-off-by: chuangcheng peng <chuangcheng.peng@amlogic.com>

action taken:
Introduce new dts parameter tee_compat. If set, the old behaviour of the
driver is enabled which solves the issue (for mecool KT1).